Intelligent Assistant/Services/UpdateData/requirements.txt
influxdb_client==1.36.0
pandas==1.3.5
schedule==1.1.0
PyYAML==6.0